Update parity check response page to support JSON body filtering
We want to be able to include/exclude keys from a JSON response when performing
a diff, so we can more easily explore the differences.

Add filter partial that renders a tree structure of checkboxes matching the
union of JSON keys from both payloads.

On checking/unchecking a checkbox the diff will refresh and include/omit the
appropriate keys.