Fix N+1 query on AB teachers page
Following on from #849 (and in a similar vein to #843), this introduces a couple
of associations to the `Teacher` model for their first and last induction
periods.

Before, we would retrieve these to return various pieces of information (e.g
induction start date) and present them to the User. This worked, but resulted in
N+1 queries since we iterate a collection of teachers.

By defining these relationships as associations, we can lean on Active Record's
automatic pre-loading to prevent them.