Introduce Mission Control
This introduces the [mission control] gem from Rails so that we can use the
lightweight dashboard to see more about the status of our background jobs.

I tried to keep the setup straight-forward, while still leaning on patterns we
have been using.

The dashboard can be enabled/disabled on a per-environment basis using
environment variables.

Admins are the only users who can access the dashboard, and they should
authenticate in the usual way. This adds some complexity (the default is HTTP
basic auth), but I think it's worthwhile in the interests of security.

Closes #1929.

[mission control]: https://github.com/rails/mission_control-jobs