Move enhancement from db:test:prepare for db:test:purge
Purge is one of the steps run by prepare. Running the command
immediately after purging means it's run when the database is freshly
recreated and no schema has been loaded or migrations run.

We want this over running before db:test:prepare because the database is
dropped and recreated after it's been enhanced.

https://github.com/rails/rails/blob/main/activerecord/lib/active_record/railties/databases.rake#L572