Update Fake API client so it always retuns data
Previously the Fake API client adhered to the RIAB approach of raising
when data was returned which meant claiming couldn't continue. The
school reg journey instead relies on conditional logic around the data
in the response.

To cope with both, the Fake API client now returns data which will cause
RIAB to raise appropriately *and* allows the school reg logic to
function.