Describe event transitions consistently
Before, we described event modifications one way, but described some event
headings another way.

At first we thought this was due to a bug #761, but in fact it was a consequence
of this inconsistency.

This introduces an `Events::TransitionDescription` service that returns a
human-readable description of an event transition. It gracefully handles blank
values like the existing code, as well as handling some extra edge cases.

Now, we can use this to describe event modifications and for event headings when
reporting events. Event timelines should now be clearer in all circumstances!