Add env-specific entries to queue.yml
In production we're seeing OOMKilled errors on workers when under load.

We suspect that the second worker in the default settings is causing
each worker to have half the memory. The migration env currently has 2GB
but production only has 1GB (it does far less, for now)

This change gives the migration env a parity-check spec second worker
but leaves production with just one multipurpose one.