Use dates instead of days for backup files
The infra team recently made a change to the storage account where
backup files can't be deleted/overwritten for a certain period of time,
after which they'll be automatically deleted.

We hit a problem today where the `Wednesday` backup couldn't be
overwritten so the backup failed.

This change switches the format so instead of being called:
  `cpdec2_production_nightly_copy_Wednesday`
it'll be called:
  `cpdec2_production_nightly_copy_2025-07-09`