Limit inialising migration db to development and migration
Dfe analytics when being instantiated does db connection checks
for models and comparisons. If this isn't disabled, when attempting
to run `ActiveRecord::Base.descendants` it will not find the db
and fall over. This means we cannot send active model entity changes
to BQ.

We can see this as we get the error `No database connection; DfE Analytics
not initialized` otherwise.