Implement basic fake begin/pass/fail/reset behaviour
We do some operations which write to the TRS API, namely when recording
induction information. The fake client should mimic the real one and
'record' updates.

This change adds that behaviour by saving the TRS values in Redis in the
format.

```
'1234567:induction' => { 'status' => ...
                         'startDate' => ...
                         'completedDate' => ...
                         'modifiedOn' => ... }
```

It also adjusts the teacher_params method so it only gives a random name
for a teacher who doesn't exist in the database, otherwise every time
the TRS refresh runs each teacher gets a new name 🎭️