Configure `:net_http` as default Faraday adapter (again)
In #782, we adopted the `:async_http` Faraday adapter so we could make parity
check requests in parallel.

As part of this change, we changed the default Faraday adapter for all clients.

Unfortunately not all ISPs support IPv6, so this broke the application for some
people (me).

To resolve this, I've configured the default Faraday adapter *back* to
`:net_http`.

The `ParityCheck::Client` explicitly sets the Faraday adapter to `:async_http`,
so the behaviour there should be unchanged.