Fix flaky test caused by dynamic generation of year
The flakiness was caused by this line:

`let(:statement_year) { statement.active_lead_provider.contract_period.year }`

The indirectly relied on the statement factory, which created an `active_lead_provider`, which created a `contract_period`, which dynamically generated a year (or reused an existing one via `.find_by_or_create_by(year:)`.

This could result in `statement_year` matching one of the contract periods in the test, leading to duplicate years in the expectation.

This fix hard explicitly sets the statement contract period year to 2034 ensuring uniqueness and removing the flakiness.