Move `TRS::TestAPIClient` specs
In #912, we split out the fake API into a fake API client and a test API client.

With that came some tests!

Those tests lived next to the code in `spec/support/trs`, but that meant we ran
those tests *every time* we ran any tests!

This moves the tests to `spec/testing`.