[1696] Show EOI choice if no confirmed partnership
If there is only an `expression_of_interest` (EOI) on the most recent `training_period` and no `school_partnership` we will show the name of the `lead_provider` from the EOI and display the fallback text in the DP section.

If the partnership is confirmed we will show the name and DP as usual.