Admins can add induction periods for Teachers without data
Before, Admins could only add an induction period for a Teacher if they already
had some induction period data.

This meant support were unable to add induction periods for "empty" Teachers.
"Empty" Teachers can happen when the only induction period is deleted or they
are imported from TRS.

This adds a link to the teacher page when there are no induction periods.

As part of this change, a bug was fixed where the induction period form was
using traning programme choices instead of induction period choices. It looks
like this was introduced in #765 (specifically
3805cb8e3d41047dcac201cb61b8da6c116cd03e).