[1940] Implement SO for identifying the contract period for a given ECT
This uses the `started_on` date from the `ECTAtSchoolPeriod` to identify the correct `ContractPeriod` for the ECT. If no match is found, it falls back to the `created_at` timestamp, which acts as the registration date.

We use the `ContractPeriod.ongoing_on(date)` scope, which checks whether a given date falls within a contract periods date range. This range is calculated by a virtual column (range) using`started_on..finished_on`.