Touch `School#api_updated_at` when school partnership changes
We want to touch the `api_updated_at` when the school partnership significantly
changes, resulting in a different `in_partnership` value returned from the
serializer.

When the first partnership is added to a school the `in_partnership` will
change from `false` to `true`.

When the last partnership is removed from a school the `in_partnership` will
change from `true` to `false`.

Supporting generic conditionals in the `DeclarativeTouch` concern and the
create/destroy scenario was going to make it very complicated, so I've ofted to
write these hooks explicitly.