Touch `School#api_updated_at` when `induction_programme_choice` changes
We want to touch the `School#api_updated_at` when the
`induction_programme_choice` of the school returned in the API serialisation
changes.

The logic for when this happens is quite complex; we determine the
`induction_programme_choice` of the school by the ECTs and mentors associated
with it:

```
- 0 ects or mentors:
  - `not_yet_known`
- 1+ ects/mentors:
  - 1+ mentor
    - `provider_led`
  - 1+ ECTs and at least 1 `provider_led` ECT:
    - `provider_led`
  - 1+ ECTs and only `school_led` ECTs:
    - `school_led`
```

This results in the following change scenarios:

| # | Event Description                                                                                          | Before         | After          |
|---|------------------------------------------------------------------------------------------------------------|----------------|----------------|
| 1 | First mentor or ECT is added to a school with none                                                         | `not_yet_known` | `provider_led` (if mentor or provider-led ECT)<br>`school_led` (if only school-led ECT) |
| 2 | A mentor is added to a school that had only school-led ECTs and no existing mentors                        | `school_led`   | `provider_led` |
| 3 | The last mentor is removed, leaving only school-led ECTs                                                   | `provider_led` | `school_led`   |
| 4 | The last provider-led ECT is removed and there are no mentors                                              | `provider_led` | `school_led`   |
| 5 | The last provider-led ECT is changed to `school_led` and there are no mentors                              | `provider_led` | `school_led`   |
| 6 | All ECTs and mentors are removed                                                                           | `provider_led` or `school_led` | `not_yet_known` |

Add `after_commit` callbacks to touch the `School#api_updated_at` when the
`induction_programme_choice` of the school changes state.

It may be worth calculating the `induction_programme_choice` before/after the
commit instead of explicitly defining these conditions, however that service is
not ready yet and may end up being more expensive to call. The tests we have
here should also be useful when if/when we switch to using a service as they
capture the minimum change definitions explicitly.