Touch `School#api_updated_at` when expression of interest changes
When the `expression_of_interest` state of a school changes (as served up by
thee schools API) we want to touch the `api_updated_at` of the school so that
it will be reflected when filtering by `updated_since`.

Touch the `School#api_updated_at` when the first expression of interest is
added for a school, or when the last expression of interest is removed (causing
it to go from `false` -> `true` and `true -> `false`, respectively).