Make ECT and mentor reg save training periods
Now, when recording either an school-led or provider-led training we
always want to write a training period.

This is part of the move towards shifting training_programme from
ECTAtSchoolPeriod to TrainingPeriod.

Training periods that are school-led won't have a school_partnership or
an expression of interest, so extra, descriptive initializers were added
to TrainingPeriods::Create to make it clear what's required for each.