Update partnership logic for touching `School#api_updated_at`
We want to touch the school's `api_updated_at` when any partnership is
created/destroyed instead of just the first/last.

This is due to the contract period scoping on the endpoint; we'd rather
over-touch the timestamp than under-touch it (which can happen if a partnership
is added for a new contract period and there are already partnerships against
the school for other periods).

Update `DeclarativeTouch` concern to support touching on specific events
(create/update/destroy).