Update induction programme logic for touching `School#api_updated_at`
We want to touch the school's `api_updated_at` when any ECT/mentor period is
created/destroyed instead of only specific conditions that would lead to the
induction programme returned from the schools API changing.

This is due to the contract period scoping on the endpoint; we'd rather
over-touch the timestamp than under-touch it (which can happen if an ECT/mentor
is added for a new contract period, depending on the state of other
participants in the same period.