Always assign an induction programme to Induction Periods
We validate an `InductionPeriod` has an `induction_programme` and we enforce
this at the database level with a non-null constraint.

Before #765 this wasn't an issue, because we always assigned an induction
programme.

Since then, we have only been assigning an induction programme if
`ENABLE_BULK_CLAIM` is false. Otherwise, we have been assigning only a
`training_programme`.

This results in validation errors wherever we try to create a new
`InductionPeriod` (e.g Admins, ABs).

Not good!

This overrides the `training_programme` setter to ensure that even when
`ENABLE_BULK_CLAIM` is true we still assign an `induction_programme`. We are
using the `PROGRAMME_MAPPER` to do this.

Now, users can create induction periods with the flag enabled and disabled
without error.

This is one approach, but I'm not sure whether it makes the most sense or not.

Alternatively we could only validate `induction_programme` when the flag is
disabled. To do this we would also need to remove the database constraint.