Enable `maxlength` for date field
We've seen a few handfuls of [errors] in production when ABs enter the date of
birth to claim an ECT.

These errors seem to happen because the user enters (or pastes?) the whole date
of birth in the day input, or these just enter too many digits.

Since these errors are not handled, the UX is degraded.

Enabling this option should prevent users entering too many digits in a given
input on this page.

We might want to do some work to handle
`ActiveRecord::MultiparameterAssignmentErrors` generally when the argument is
out of range, but we can cover that in another PR.

[errors]: https://dfe-teacher-services.sentry.io/issues/6385934864/?environment=production&project=4508369974788096&query=&referrer=issue-stream