Add `Metadata::SchoolLeadProviderContractPeriod`
Add a model to encapsulate metadata about a school/lead provider/contract
period relationship.

This is school-level data/an aggregation of scoped participant data relative to
the school.

Namespaced under `Metadata` to ensure its not used inappropriately; it is also
locked down so that only the `Metadata` namespace/services can updated the
metadata models.