Ensure lead providers are present and have an `ecf_id`
This is just a small housekeeping update to avoid running into a silent failure
when trying to run the parity check when there are no lead providers that have
an `ecf_id` (it would hang/fail trying to get tokens).

Prevent running the migrator until there are valid lead providers present.

Only queue requests for lead providers with an `ecf_id`.