Fix `not_cip_only` school factory trait
The scopes in `GIASHelpers` appear to check the `funding_eligibility` of the
`GIASSchool` when determining if it is cip-only or not-cip-only. The factory
trait, however, sets the `type` as CIP only or not.

Update the factory trait to set the expected `funding_eligibility` for
`cip_only` and `not_cip_only`.