Add `api_updated_at` to `DeliveryPartner`
We want to surface the delivery partners relative to a given `updated_since` in
the API, but only when certain fields are changed (the ones that we return,
serialized).

Add an `api_updated_at` attribute to `DeliveryPartner` that we can touch when a
field we expose over the API is changed (this may be on a related model, not
just `DeliveryPartner`).