Touch `DeliveryPartner#api_updated_at` on relevant changes
We want to update the `api_updated_at` of a `DeliveryPartner` when a relevant
attribute changes that would result in the serialized response being different.

Looking at ECF, for a delivery partner this would be when the `name` of the
delivery partner changes, or its relevant `cohort(s)`.

Touch `api_updated_at` when the delivery partner `name` changes or when a new
`LeadProviderDeliveryPartnership` is added/removed.

Assumes lead provider delivery partnerships are not expected to change their
`active_lead_provider` or `delivery_partner` and `ActiveLeadProvider` would
never change its `contract_period`.