Add `DeliveryPartnerSerializer`
Add the delivery partner serializer. We return the `name` of the delivery
partner, the `created_at` and `api_updated_at` and the `cohort`, which is
retrieved via the `lead_provider_delivery_partnerships`.

We do not need to `uniq` the cohorts as there is already a unique index on
`active_lead_provider` and `delivery_partner` for
`lead_provider_delivery_partnerships` and also on `lead_provider_id` and
`contract_period_id` for `active_lead_providers`.