Add date relative scopes to interval
While working on moving the training programme from ECTAtSchoolPeriod to
TrainingPeriod[0] I realised some work needed to be done on the query
objects and scopes to support it.

When viewed by a school the ECTs index should list teachers who are:

* currently ECTs at the school
* scheduled to join the school (i.e., their started_on date is in the
  future)

Two new scopes were added and a third, which chains those two together,
is used by the Teachers::Search class.

[0] DFE-Digital/register-ects-project-board#2013