Add seed data so we can preview future joiners
The new teacher from the future, Peter Davison, will be used to
demonstrate how teachers who've not yet joined a school but are
scheduled to.

The seed output looks like this:

🌱   Peter Davison (ECT)
🌱     * has been an ECT at Abbey Grove School from 2025-08-11 (ECT at school period)
🌱     * will be trained by Teach Fast (LP) and Grain Teaching School Hub from 2025-08-11 (training period)

The seeds generally aren't dynamic enough so there's a point where 2025
won't be the current cohort, we might need to make them all relative
rather than hardcoding years.