Add `DeliveryPartners::Query`
Add a query service for retrieving delivery partners, which will be called from
the controller. TThe results will be passed into the
`DeliveryPartnerSerializer`.

We calculate a `transient_cohort` in the query service for efficiency and
fallback to looking it up from sratch if its not present. This will be replaced
in the future by a precalculated field.