Fix bug where names starting with titles are stripped
Occasionally we have a failing spec like this:

  1) Migrators::Teacher behaves like a migrator #migrate! creates a Teacher records for each ECF TeacherProfile
     Failure/Error: expect(Teachers::Name.new(teacher).full_name).to eq user.full_name

       expected: "Missy Sippy"
            got: "y Sippy"

It's because the full name parser strips 'Miss' from the first name (to
protect against occasions where a user has accidentally included the
title in their first name field).

Some names, i.e. Missy, start with a title. Now the mastcher looks for a
space after the title.