Add support for different parity check URLs
On some of the new RECT endpoints we are omitting `ecf` from the path of the
endpoint (schools, for example). This means we need to call different endpoints
depending on the application.

Add support to parity check endpoint options for `ecf_url` or `rect_url` - this
will be used with preference over the default URL provided as the key to the
check.