Remove cancel-in-progress from deploy workflow
This has been added for concurrency however causes issues with when the
job is on the terraform apply step and then the job is cancelled.

Opt to have current jobs complete before kicking off the next job to
avoid getting into strange states where resources created are not
confirmed in state.