Pass relevant objects into listing card component
Rather than pass in just the ECTAtSchoolPeriod and then traverse using
logic inside the component, we should pass in objects and use them
directly.

This keeps the logic of _what_ to display out of the components (which
should be concerned with _how_ to display stuff). Now the
traversal happens in the view, which isn't ideal, but this can easily be
moved.

Plus, this makes testing much easier as objects can be created in
isolation.