Rework ECTAtSchoolPeriods::Training
This changes the class so it doesn't differentiate between the latest
and current training period. It's needed because we now allow future
periods to be recorded and we're look at training from a school's point of
view, via the ECTAtSchoolPeriod. This means we'll be looking at:

* either the current (ongoing now) training period
* a training period due to start in the future

There's no need (at the moment) for future training periods at the same
school, so we only want to retrieve one.