Only support filtering on `index` actions
Currently the filtering will also be applied to the `show` action.

Restrict the query conditions so that filters only get passed in on `index`
actions (apart from filtering on `contract_period_id` for the schools `show`
action as we allow this).

Add minimum tests to ensure `show` actions do not allow filtering.