Support response bodies with arrays of values
Currently we expect all arrays to contain objects (which can be mapped into a
hash). When we return `cohort` for delivery partners, however, it is an array
of strings and trips up the `ParityCheck::Filter::ResponseBody` logic.

Ignore nested items in the key hash mapping unless they are also a hash.