Ignore `Rails/SaveBang` on `FactoryBot` usage
Rubocop is failing due to the `FactoryBot.create` call in
`metadata_support.rb`. I'm not sure why this is being picked up as an issue in
this file but not elsewhere; maybe its to do with it not being in a `_spec.rb`
file, but either way we should ignore it.