Use Teacher `last_induction_period` association
In #858 we introduced a `last_induction_period` association to the `Teacher`
model to fix some N+1 queries.

Now this association exists, it felt simpler to lean on this rather than
accessing it through the service object.

This moves the `last_induction_period` method out of the public API for the
`Teachers::InductionPeriod` service and updates usage in a few places.