Admins can only "Reopen induction" when there is an outcome
In #522 we introduced functionality that allowed Admin users to reopen a
completed induction.

This works well, but it also allows Admin users to "reopen" a completed
induction without an outcome (i.e the induction is "paused"). This was not the
original intention.

Even though this works without issue, it adds a level of complexity for Admin
users that isn't necessary. Since they can achieve the desired outcome already
by "Edit"ing a paused induction, we can safely remove this.

Now, Admin users can only "reopen induction" when the latest induction period is
completed **and** has an outcome (i.e "pass" or "fail").