Retry background jobs when the TRS API errors
As part of [this investigation] into mismatched TRS induction statuses, we
identified that our background jobs are not retrying by default.

That meant that any intermittent API error (e.g service unavailability) would
result in some jobs failing, and the corresponding updates *never happening*.

Now, in jobs that use the TRS API we retry on API request errors. As part of
this change, the `Teachers::RefreshTRSAttributes` service has been refactored so
we can use a single `api_client` in a job, rather than instantiating several.

[this investigation]: https://github.com/DFE-Digital/register-ects-project-board/issues/2134