Filter metadata in `Schools::Query`
We only care about the metadata associated with the given
`contract_period_year` and `lead_provider_id` passed into
the schools query.

Filter the to-many relationship to only return relevant
metadata. Using `references` provides an `LEFT OUTER JOIN`
on the metadata tables, which we can then filter. We include
`nil` to ensure schools are returned even if there are no
metadata records (which should never be the case, but feels
more robust in case it ever happens and also means the other
test setup is leaner).

We could update the serializer to not pass in the
`lead_provider_id` and `contract_period_year`, but it feels
safer to do this so that the serializer isn't tightly
coupled to the query.