Admins must confirm before reopening an induction
Before, reopening a Teacher's induction was a one-click action.

We've seen some mistakes since launching this feature, so we think we've made it
too easy to use accidentally. Especially considering the action is destructive
and irreversible.

This adds some friction to the journey by introducing a confirmation screen.

Now, admins must confirm they are sure before then can reopen a Teacher's
induction.

As part of this change, we are "refreshing" the TRS data after we've reopened
the induction with them. This follows the same pattern adopted by the pass/fail
induction journeys.