Handle future contract periods
If the contract period is not yet open and has never previously been open,
there will be a list of available lead providers as checkboxes.
Only lead providers who are active for the contract period should be shown
in the list of options. Some of the LPs will already be checked if they
have been selected before, and can be unchecked as needed.