Fix flakey parity check request spec
I suspect as our IDs increment, depending on the test order, we will have runs
where the ID is generated as `99`, which is what we use as our 'not found'
example. This will cause the response to succeed and the test to fail (which is
what we see in the CI logs - it reports 200 instead of 404). Using `-1` as the
ID should ensure it will always fail to find the resource.