Switch to using real lead provider names
Originally we used similar but different lead provider names in development and
review. This was in an effort to reduce the risk of changing something in an
environment accidently, however its caused more issues and confusion than we
expected so we are going to use the normal/real lead provider names everywhere.