Add `failed` state to `ParityCheck::Run` and `ParityCheck::Request`
We want to move a parity check run into a failed state if it encounters errors,
instead of having the jobs fail which leads to the run being stuck as in
progress.

Add `failed` states to parity check `Request` and `Run` models.

I've opted to use `halt` as the method that fails a request/run as both `fail`
and `abort` are reserved in Ruby.