Move `Run` and `Request` to `failed` state on error
When the parity check encounters an error we want to move the `Request` that
errors and, ultimately, the whole `Run` to a `failed` state.

Clean up error handling for easier rescues.

Rescue `Client::Error` in the `RequestHandler` and fail the request.

Check for failed requests in the `RequestDispatcher` and move the `Run` to
`failed` if any of the requests have failed.

Note there are two errors I am not rescuing; `RunModeError` and
`UnsupportedEnvironmentError` -- if these are hit then something is very wrong
with the UI or environment.