Fix flakey school partnership touch spec
I can't see why its flakey, but we have another test that is similar but
uses a different format for the new value (it creates the associated
object with the generated ID value instead of overriding the
`generate_new_value` method).

Changing this spec to be inline with that one to see if it resolves the
flakiness.