Admins must add context when reopening an induction
In #1133 we added some friction to the journey when Admin users reopen a Teacher
induction.

The intention of this was to reduce the number of mistakenly reopened
inductions.

This is an improvement, but we also want Admin users to be able to add more
context about **why** they are reopening an induction.

Now, they must add a note or Zendesk ticket ID when they reopen an induction.
This information is saved on the event, and is displayed in the Teacher
timeline.

This leans on some work originally proposed in #251.