Add method to select relevant lead provider
When there is both an expression of interest and a school partnership
attached to a training period, we want to favour the lead provider
attached to the school partnership over the one attached to the
expression of interest.

In most cases they'll be the same, but (probably) not always.

This change adds #lead_provider_via_school_partnership_or_eoi which
implements the logic along with tests.