Allow running ERD diagrams in test env
We are going to do this on a nightly job. As the CI is setup to work with
the `test` environment we need to allow this for ERD generation as well (I
tried moving CI to use `development` for the action but there are other,
harder to solve issues in doing that).