Add API error mapping service
We need to be able to swap out the RECT terminology used in the service
errors and attribute names with ECF terminology that lead providers are
familiar with.

Add `API::Errors::Mapper` service that will transform RECT terms to ECF.

Hook into the `API::Errors::Response` service to replace error messages.