Add partnerships POST/PUT endpoints to parity check
Add the POST and PUT endpoints to the parity check endpoints config file.

Update the `DynamicRequestContent` with the relevant request body methods.

Fallback to returning a `nil` body if required data can't be found to make a
valid request; this will result in a 422 in the parity check, which should be
fine and will prevent it from just falling over.