Add MentorshipPeriods::Finish service
When an is marked as leaving a school (past, present or future) their
mentorship cannot continue beyond the point they're at the school, so we
must close the mentorship periods on the day their ECT at school period
ends.

There will probably be future cases where mentorships end for other
reasons too, like if the mentor becomes unavailable or a more suitable
person emerges.

The service closes the mentorship period and records events that will
be visible on both the mentor and mentee's timelines.

Currently it's only called when closing the related ECTAtSchoolPeriod.