Optimise query services and serialization
Currently our query services are responsible for preloading the relevant
relationships that we reference during serialisation. This has two issues:

1. It's slow to preload and paginate
2. Queries should be generic/not coupled to a particular use case

To remedy this, the commit pulls out any unnecessary `includes` from the query
services and instead passes the output from a query to a class method on the
relevant serializer to perform specific optimisations.

This means we only do the heavy lifting on the result of the pagination and
also places the optimisations relevant to the serializer in the serializer
itself, so its clearer when it should need to change.

As part of this change we need to move back to `pagy` instead of
`pagy_countless`. This is because `pagy_countless` internally calls `to_a`,
unlike `pagy` which returns the ActiveRecord collection. We could get around
this by plucking the `id` in the `pagy_countless` result and then re-querying,
but in my initial testing this doesn't make much of a performance difference.