Optimise query services and serialization
Currently our query services are responsible for preloading the relevant
relationships that we reference during serialisation. This has two issues:

1. It's slow to preload and paginate
2. Queries should be generic/not coupled to a particular use case

To remedy this, we are not preloading on the base query and instead only
preload/include on paginated results.

As part of this change we need to move back to `pagy` instead of
`pagy_countless`. This is because `pagy_countless` internally calls `to_a`,
unlike `pagy` which returns the ActiveRecord collection. We could get around
this by plucking the `id` in the `pagy_countless` result and then re-querying,
but in my initial testing this doesn't make much of a performance difference.

Add shared context to fail a test if an includes is added to a base query. We
want to keep these optimal; any includes should be added on the returned scope.

Currently, if we reference a new field in a serializer its easy to forget to
eager load the relationship.

Add strict_loading to the API queries to ensure N+1's are avoided. Originally I
was going to add this to the base queries but hit failures outside of the API
so opted for a safer approach.