Schools ECT page renders even when there is no training period
Before, visiting a schools ECT page where there was no current training period
would result in an error.

Since it is possible for an ECT to have no training period, this seemed like a
reasonable edge case we should handle.

This updates the `Schools::ECTTrainingDetailsComponent` so it only renders when
the training period is present.

As part of this change, I have introduced a `have_summary_list_row` matcher,
which I think improves the ergonomics of asserting against key <> value pairs in
summary lists.