Ensure serializers perform eager loading
Currently, if we reference a new field in a serializer its easy to forget to
eager load the relationship.

Add `strict_loading` to the API queries to ensure N+1's are avoided. Originally
I was going to add this to the base queries but hit failures outside of the API
so opted for a safer approach.