Add ImpersionationController and route
The ImpersionationController has two actions:

* create  - to start impersonating
* destroy - to finish impersonating

They allow DfE users to impersonate school users by modifying the
session so there's a new user type and school_urn.

The functionality to build a session is in the ImpersonateSchoolUser
module, as it's needed by both the DfEPersona and DfEUser.

The functionality to revert the session is just in
DfEUserImpersonatingSchoolUser.