Schools can change an ECT's email address
This introduces a wizard for updating an ECT's email address.

Updates are stored in the session until the User confirms the change (after
which the ECT record is updated).

This is one of many attributes schools will be able to be update about ECTs,
which will ensure any errors or changes can be made autonomously by end users
via the Teacher details page.

In the spirit of "make the change easy, then make the easy change", this
introduces a `Wizardable` concern which handles lots of the boilerplate for
plugging in a mini-wizard controller to allow Schools to make changes to ECTs.